Simplified kotus2json.py (and a bug fix; TN).

+++ kotus2json.py
@@ -1,5 +1,7 @@
 #!/usr/bin/env python3
-'''kotus2json.py -- convert Kotus XML to JSON
+'''
+
+  kotus2json.py -- convert Kotus XML to JSON
 
   Usage:
       kotus2json.py xmlfile [> jsonfile]
@@ -10,6 +12,8 @@
   2017-10-06  0.4.0  Upload to GitHub. (TN)
   2017-10-09  0.4.1  Bug fix. I had changed KotusParser to handle a stream,
                      not a file. File has to opened here. (TN)
+  2017-10-26  0.5.0  Simplified structure. A list will suffice for each
+                     lexeme instead of a dict.
 
 '''
 
@@ -18,30 +22,30 @@ import os.path
 import json
 import kotusparser
 
-version = '0.4.1'
+version = '0.5.0'
 
 def die(msg):
-    print('{}: {}'.format(os.path.basename(sys.argv[0]), msg), \
-          file=sys.stderr)
+    prg = os.path.basename(sys.argv[0])
+    print('{}: {}'.format(prg, msg), file=sys.stderr)
     sys.exit(1)
 
 if not sys.argv[1:]:
     die('Usage: kotus2json XMLFILE...')
-
 for arg in sys.argv[1:]:
+    if not os.path.exists(arg):
+        die('Not found: {}'.format(arg))
     jsonfile = os.path.splitext(arg)[0] + '.json'
     vocab = []
     try:
         with open(arg, 'r') as xml:
             try:
-                for word, para in kotusparser.KotusParser(xml):
+                for word, para, grad in kotusparser.KotusParser(xml):
                     if word:
-                        lex = {'s': word, 'tn': para if para else 0}
-                        vocab.append(lex)
+                        vocab.append([word, para, grad])
             except kotusparser.ParseError:
                 die('Cannot read or parse error: {}'.format(arg))
-    except (FileNotFoundError, PermissionError, IOError):
-        die("File not found, no read permission or I/O error: {]".format(arg))
+    except (PermissionError, IOError):
+        die("Cannot read: {]".format(arg))
     try:
         with open(jsonfile, 'w') as f:
             json.dump(vocab, f, ensure_ascii=False, indent=1)
